ppc::util::GetPPCNumThreads()

+++ tasks/stl/voroshilov_v_convex_hull_components/src/chc.cpp
@@ -1,4 +1,5 @@
 #include "../include/chc.hpp"
+#include "core/util/include/util.hpp"
 
 #include <algorithm>
 #include <atomic>
@@ -23,7 +24,7 @@ Image::Image(int hght, int wdth, std::vector<int> pxls) {
   width = wdth;
   pixels.resize(height * width);
 
-  int num_threads = static_cast<int>(ppc::util::GetPPCNumThreads());
+  int num_threads = ppc::util::GetPPCNumThreads();
   int chunk = (height + num_threads - 1) / num_threads;
   std::vector<std::thread> threads;
 
@@ -185,7 +186,7 @@ std::vector<Component> voroshilov_v_convex_hull_components_stl::FindComponentsST
   int height = tmp_image.height;
 
   std::vector<std::thread> threads;
-  int num_threads = static_cast<int>(ppc::util::GetPPCNumThreads());
+  int num_threads = ppc::util::GetPPCNumThreads();
   std::vector<std::vector<Component>> local_components(num_threads);
   int chunk_height = (height + num_threads - 1) / num_threads;
   std::vector<int> y2(num_threads);
@@ -321,7 +322,7 @@ std::vector<Hull> voroshilov_v_convex_hull_components_stl::QuickHullAllSTL(std::
   std::vector<Hull> hulls(components.size());
 
   std::vector<std::thread> threads;
-  size_t num_threads = ppc::util::GetPPCNumThreads();
+  size_t num_threads = static_cast<size_t>(ppc::util::GetPPCNumThreads());
   size_t chunk = (components.size() + num_threads - 1) / num_threads;
 
   for (size_t t = 0; t < num_threads; t++) {
@@ -351,7 +352,7 @@ std::pair<std::vector<int>, std::vector<int>> voroshilov_v_convex_hull_component
   std::atomic<int> uniq_hull_index(1);
 
   std::vector<std::thread> threads;
-  size_t num_threads = ppc::util::GetPPCNumThreads();
+  size_t num_threads = static_cast<size_t>(ppc::util::GetPPCNumThreads());
   size_t chunk = (hulls.size() + num_threads - 1) / num_threads;
 
   for (size_t t = 0; t < num_threads; t++) {
